fix(oddshap): correct Fourier design matrix sign handling
  Cast the parity matrix to float before applying the Fourier sign transform.
  This prevents uint8 underflow where -1 became 255 and restores full-budget
  consistency against ExactComputer. Also removes obsolete xfail markers for
  the fallback and full-budget convergence tests.

+++ src/shapiq/approximator/regression/oddshap.py
@@ -523,7 +523,9 @@ class OddSHAP(Approximator):
         interaction_masks_int = interaction_masks.astype(np.uint8)
 
         parity_matrix = (coalitions_int @ interaction_masks_int.T) % 2
-        design_matrix = 1 - 2 * parity_matrix
+        # Cast before the Fourier sign transform: uint8 would underflow
+        # 1 - 2 * 1 to 255 instead of -1.
+        design_matrix = 1.0 - 2.0 * parity_matrix.astype(float)
         X_tilde = design_matrix.astype(float) * row_weights_used[:, np.newaxis]
         y_tilde = centered_values * row_weights_used
